Last commit: End of the term
It's been a journey (a short one, honestly) since I first started using Git. I started learning it in the middle of the scholar year, but since then I got used to it quickly, so that's the reason why the first commit had lots of terms at once. I believe (and hope) this will be the last commmit because I already ended the final exam, so if you don't see any new update here, that's because I've passed it. Yay!